Began restructuring Added basic service and repository files
