+++ Habitual/Connection.swift
@@ -90,66 +90,4 @@ class Connection: ParseObject {
         
         loadMessages(nil)
     }
-    
-    /// The method which which all messages should be sent. A message is
-    /// created with current connection (with the current `User` as a sender), 
-    /// sent to the server, appended to `messages` and then returned.
-    ///
-    /// - parameter text: The text of the message to send.
-    ///
-    /// - returns: The newly created `Message` that has been asynchronously sent
-    ///   to the server.
-    func sendMessage(text: String) -> Message {
-        let message = Message(text: text, connection: self)
-        message.send()
-        messages?.append(message)
-        return message
-    }
-    
-    /// Load every `Message` associated with this connection from the server and
-    /// replace the content of `messages` with the results.
-    ///
-    /// - parameter callback: A closure which is called with `success`, a `Bool`
-    ///   describing whether the API call was successful and the messages were
-    ///   loaded.
-    func loadMessages(callback:((success: Bool) -> ())?) {
-        let query = PFQuery(className: "Message")
-        query.whereKey("connection", equalTo: self.parseObject!)
-        query.orderByAscending("timeStamp")
-        query.includeKey("sender")
-        query.includeKey("habit")
-        query.findObjectsInBackgroundWithBlock { (parseObjects, error) -> Void in
-            if let parseObjects = parseObjects {
-                var messages: [Message] = []
-                for parseObject in parseObjects {
-                    let message = Message(parseObject: parseObject)
-                    messages.append(message)
-                }
-                
-                self.messages = messages
-                
-                if let callback = callback {
-                    callback(success: true)
-                }
-            }
-        }
-    }
-
-    /// Save the underlying 'PFObject' to the server.
-    func saveToServer() {
-        parseObject!.saveInBackground()
-    }
-    
-    /// Approve the `Connection`. The approved `Connection` is then saved to the
-    /// server.
-    func approve() {
-        parseObject!["privateApproved"] = true
-        approved = true
-        parseObject!.saveInBackground()
-    }
-    
-    /// Delete the 'Connection' from the server.
-    func deleteFromServer() {
-        parseObject!.deleteInBackground()
-    }
 }
